Revise argument limit from 2^16 to 2^8
This means calls can have a single byte operand instead of an awkward two-byte operand for a 3-byte instruction. Let's be honest if a user wants a 256+ parameter function then they're doing it wrong and can just use an array anyway.

+++ src/object.h
@@ -30,7 +30,7 @@ struct object {
 
 struct object_function {
     object obj;
-    uint16_t arity;
+    uint8_t arity;
     segment seg;
     object_string *name;
 };
